Allow fallback to width/height for symbol viewBox
Introduces a new recognized option called 'whviewbox'.

Ordinarily, each generated symbol in the output file only gets a viewBox
if the original source SVG had one. When enabled, whviewbox makes it so
that having both a width and height on the original SVG will allow a
viewBox attribute to generated automatically.